% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{align.dataset.margins}
\alias{align.dataset.margins}
\title{align dataset margins}
\usage{
align.dataset.margins(datasets, na.rows = TRUE, na.cols = FALSE,
  use.dims = NULL)
}
\arguments{
\item{datasets}{list of matrices (columns are samples, rows are genes)}

\item{na.rows}{if TRUE, add NAs for missing rows, otherwise they are removed}

\item{na.cols}{if TRUE, add NAs for missing columns, otherwise they are removed}

\item{use.dims}{either an integer or character vector of length 1}
}
\value{
new list of matrices
}
\description{
align dataset margins
}
