% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cn.R
\name{cluster_cn}
\alias{cluster_cn}
\title{Hierarchically cluster copy number data}
\usage{
cluster_cn(cn_vec, min_dpts = 25, cluster_method = "average",
  n_segments = 25)
}
\arguments{
\item{cn_vec}{vector of copy number values for each genomic locus in a particular cell line}

\item{min_dpts}{sets the minimum number of data points within a cluster leaf}

\item{cluster_method}{the agglomeration method to be used. See \code{\link[stats]{hclust}} for available methods}

\item{n_segments}{the number of segments in the piecewise linear cutting effect fit}
}
\value{
A vector of knot points of the form c(start_1, end_1, start_2, end_2, start_3, end_3, ..., start_n, end_n)
}
\description{
Hierarchically cluster copy number data to create knot points for piecewise linear cutting effect fit
}
\examples{

}
