% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_prediction_auc.R
\name{evaluate_prediction_auc}
\alias{evaluate_prediction_auc}
\title{Calculate area under the curve for ROC and PR curves.}
\usage{
evaluate_prediction_auc(predicted_values, true_values, min_resolution = 1000,
  plot = T, fdr_threshold = 0.05)
}
\arguments{
\item{predicted_values}{Vector of continuous-valued prediction scores. 
Assumes that positive prediction scores correspond to 1 outcome in the true values.}

\item{true_values}{Vector of logical or binary values of ground truth}
}
\value{
A data.frame of containing the AUROC, AUPRC, and F1 measure
}
\description{
Calculate area under the curve for ROC and PR curves.
}
\examples{
true_values <- sample(0:1, 1000, replace=T)
predicted_values <- rnorm(1000)
evaluate_prediction_auc(predicted_values, true_values)
}
