% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{generate_guides}
\alias{generate_guides}
\title{Generation of the guides from a gct dep_file}
\usage{
generate_guides(dep_file, write_rds_output_path = NULL)
}
\arguments{
\item{dep_file}{file path of guide-level dependency data}

\item{write_rds_output_path}{Optional: Will write guide_dep into a rds file in top of returning the matrix}
}
\value{
Matrix with a description attribute
}
\description{
Generation of the guides from a gct dep_file
}
