% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_alignments.R
\name{guideAlignments}
\alias{guideAlignments}
\title{Read BAM file report, filter with valid PAM sequence, report # of alignments}
\usage{
guideAlignments(bam.file, max.alns = 100, pam = "[ACGTN]GG|GG[ACGTN]",
  genome_id = "hg19", chromosomes = paste0("chr", c(as.character(1:22), "X",
  "Y")), include.no.align = F, as.df = T, guide_length = 20)
}
\arguments{
\item{bam.file}{BAM file path}

\item{max.alns}{only consider guides with fewer than this many alignments}

\item{pam}{PAM sequence as regular expression}

\item{include.no.align}{include guides that do not map to genome}

\item{as.df}{return data.frame if \code{TRUE}, GRanges object if \code{FALSE}}
}
\value{
data.frame or GRanges object of results
}
\description{
Read BAM file report, filter with valid PAM sequence, report # of alignments
}
