% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_seg_intersect.R
\name{intersect_data_with_segment}
\alias{intersect_data_with_segment}
\title{Finds overlaps between target segments and data. Calculates result of a function across overlaps.}
\usage{
intersect_data_with_segment(seg, gr, dat.column, func = median,
  na.value = NA, do_parallel = F, ...)
}
\arguments{
\item{seg}{GRanges object to calculate data over}

\item{gr}{GRanges object with a data column}

\item{dat.column}{which mcols name to find data in \code{gr}}

\item{func}{what function to use (e.g.: \code{median}, \code{mean}, \code{length})}

\item{na.value}{value to use if segment has no overlaps with gr}

\item{\dots}{extra arguments passed to \code{func}}
}
\value{
a vector with one value per segment
}
\description{
Finds overlaps between target segments and data. Calculates result of a function across overlaps.
}
