% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{intersect_locus_with_cn_seg}
\alias{intersect_locus_with_cn_seg}
\title{Intersect locus with copy number segment}
\usage{
intersect_locus_with_cn_seg(cn_seg, guide_alns, cell_lines = NULL,
  genomeinfo = Seqinfo(genome = "hg19"), chromosomes = paste0("chr",
  c(as.character(1:22), "X", "Y")), do_parallel = F)
}
\description{
Intersect locus with copy number segment
}
