% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{make.a.tidy.dataset}
\alias{make.a.tidy.dataset}
\title{make a tidy dataset}
\usage{
make.a.tidy.dataset(datasets, na.rows = TRUE, na.cols = FALSE,
  use.dims = NULL, dim.names = c("Gene", "Sample"))
}
\arguments{
\item{datasets}{list of matrices (columns are samples, rows are genes)}

\item{na.rows}{if TRUE, add NAs for missing rows, otherwise they are removed}

\item{na.cols}{if TRUE, add NAs for missing columns, otherwise they are removed}

\item{use.dims}{either an integer or character vector of length 1}

\item{dim.names}{character vector of length two -
these become the column labels in the tidy dataset
for the rownames and colnames of the original matrices}
}
\value{
a tidy dataframe
}
\description{
Takes a list of matrices with similar row and column names
and runs tidyr::gather across each to make a "long" dataset
with a column of values for each matrix in the list.
}
