% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{map_guide_to_locus}
\alias{map_guide_to_locus}
\title{Map guide to locus}
\usage{
map_guide_to_locus(guides, genome_id = "hg19", bowtie_exe = "bowtie",
  samtools_exe = "samtools", temp_dir = tempdir(),
  write_rds_output_path = NULL, guide_length = 20)
}
\description{
Map guide to locus
}
