% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read.gct}
\alias{read.gct}
\title{read a *.gct file into your session}
\usage{
read.gct(gct.file)
}
\arguments{
\item{gct.file}{path to gct formatted file}
}
\value{
matrix with a "description" attribute
}
\description{
read a *.gct file into your session
}
