% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read.gmt}
\alias{read.gmt}
\title{read a *.gmt file into your session}
\usage{
read.gmt(gmt, as.df = T)
}
\arguments{
\item{as.df}{whether to return as data frame or list}

\item{gmt.file}{path to gmt formatted file}
}
\value{
either a list of character vectors or a two column data.frame
}
\description{
read a *.gmt file into your session
}
