% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_precision_recall.R
\name{run_precision_recall}
\alias{run_precision_recall}
\title{Run precision recall assessment using essential and nonessential genes}
\usage{
run_precision_recall(ge_fit)
}
\arguments{
\item{ge_fit}{Matrix of gene effects by cell line. Expects row names are gene symbols and column names are cell line IDs.}
}
\value{
List of PR results.
}
\description{
Run precision recall assessment using essential and nonessential genes
}
