% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ceres.R
\name{wrap_ceres}
\alias{wrap_ceres}
\title{CERES wrapper}
\usage{
wrap_ceres(sg_path, cn_path, guide_locus_path, locus_gene_path,
  replicate_map_path, run_id, optimize = F, params = NULL,
  cl_subset = NULL, fit_efficacy = T)
}
\arguments{
\item{sg_path}{path to matrix with sgRNA sequences as row names, samples (i.e. replicates of cell line screens) as column names}

\item{cn_path}{path to matrix with genomic cut sites as row names, cell lines as column names}

\item{guide_locus_path}{path to data.frame with column names `Guide` and `Locus`. Values must match those corresponding to row names of sg_data and cn_data}

\item{locus_gene_path}{path to data.frame with column names `Locus` and `Gene`}

\item{replicate_map_path}{path to data.frame with column names `Replicate` and `CellLine`}

\item{run_id}{name for the CERES run}

\item{optimize}{toggle optimization routine}

\item{params}{list of parameters; must specify or use defaults when optimize = F}

\item{fit_efficacy}{boolean indicating whether sgRNA offsets and activity scores are computed}
}
\value{
A list of numeric vectors and matrices containing the results of the CERES fit.
}
\description{
CERES wrapper
}
\examples{

}
