% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{write.gct}
\alias{write.gct}
\title{write a *.gct file to disk}
\usage{
write.gct(mat, gct.file, description = attr(mat, "Description"))
}
\arguments{
\item{mat}{data matrix, optionally with attribute "description"}

\item{gct.file}{path to file to write}

\item{description}{character vector; length must equal number of rows in mat}
}
\description{
write a *.gct file to disk
}
